Convert to Int32 instead of Int64

+++ LXGaming.Common.Newtonsoft/Converters/IntegerBoolConverter.cs
@@ -10,7 +10,7 @@ public class IntegerBoolConverter : JsonConverter<bool> {
 
     public override bool ReadJson(JsonReader reader, Type objectType, bool existingValue, bool hasExistingValue, JsonSerializer serializer) {
         if (reader.TokenType == JsonToken.Integer) {
-            var value = Convert.ToInt64(reader.Value);
+            var value = Convert.ToInt32(reader.Value);
             if (value is 0 or 1) {
                 return value == 1;
             }
